requirements.txt
pandas>=2.0.0
argparse
pathlib
groq>=0.4.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
mcp>=0.1.0
python-dateutil>=2.8.2
huggingface-hub>=0.19.0
transformers>=4.35.0
scikit-learn>=1.3.0
numpy>=1.24.0
